# api/dist/swagger.json
{
	"components": {
...
		"schemas": {
...
			"TransactionsBody": {
				"properties": {
					"coin": {
						"type": "string"
					},
					"startTime": {
						"type": "number",
						"format": "double"
					},
					"endTime": {
						"type": "number",
						"format": "double"
					},
					"startBlock": {
						"type": "number",
						"format": "double"
					},
					"endBlock": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"coin"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "pioneer-server",
		"version": "0.0.12",
		"description": "Pioneer is a multi-asset blockchain explorer and subscription service",
		"contact": {
			"name": "highlander ",
			"email": "bithighlander@gmail.com"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/username/{username}": {
			"get": {
				"operationId": "Username",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "get public user info",
				"tags": [
					"Public Endpoints"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "username",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/invocation/{invocationId}": {
			"get": {
				"operationId": "Invocation",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "get public user info",
				"tags": [
					"Public Endpoints"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "invocationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/invocations": {
			"get": {
				"operationId": "Invocations",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Private Endpoints"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "Authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/context": {
			"get": {
				"operationId": "Context",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Get the balances for a given username",
				"tags": [
					"Private Endpoints"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "Authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/transactions": {
			"post": {
				"operationId": "Transactions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Get transaction history\n\nstart block\nend block\ncoin",
				"tags": [
					"Private Endpoints"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "Authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TransactionsBody"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "https://pioneers.dev/api/v1"
		}
	]
}